Show user privacy settings in new dashboard
