+++ sematic/ui/packages/main/src/index.tsx
@@ -11,6 +11,7 @@ import PipelineRuns from "@sematic/common/src/pages/PipelineRuns";
 import NewRunDetails from "@sematic/common/src/pages/RunDetails";
 import NewRunSearch from "@sematic/common/src/pages/RunSearch";
 import { getFeatureFlagValue } from "@sematic/common/src/utils/FeatureFlagManager";
+import { setupPostHogOptout } from "@sematic/common/src/utils/postHogManager";
 import { ExtractContextType } from "@sematic/common/src/utils/typings";
 import { useAtom } from "jotai";
 import { RESET } from "jotai/utils";
@@ -35,7 +36,6 @@ import LoginPage from "./login";
 import PipelineIndex from "./pipelines/PipelineIndex";
 import RunView from "./pipelines/PipelineRunView";
 import PipelineView from "./pipelines/PipelineView";
-import { setupPostHogOptout } from "./postHogManager";
 import { RunIndex } from "./runs/RunIndex";
 import { sha1 } from "./utils";
